app/requirements.txt
Flask==2.2.2
openai[embeddings]==0.27.2
pinecone_client==2.0.13
requests_html==0.10.0
validators==0.20.0
gunicorn==20.1.0